Allow using system installed AFLPlusPlus
While it would be preferably to always use zig built AFLPlusPlus, this does not consistently work.
On some systems (like ubuntu github runners), llvm uses C++11 abi for some functions.
This leads to the functions having `B5cxx11` in the name.
When zig compiles AFLPlusPlus, it misses this part of the name mangling.
As such, the zig compile objects have different symbol names and failed to complete linking.

Since I was unable to resolve the above issue, I decided to enable avoiding zig compiled AFLPlusPlus.
Instead, when the workaround is enabled, it will use `afl-cc` found in PATH.